.Refactor cleanup
