+++ test/integration/helper/constants.go
@@ -1,0 +1,6 @@
+package helper
+
+const (
+	SokarAddr string = "http://localhost:11000"
+	NomadAddr string = "http://localhost:4646"
+)
