+++ test/integration/main_test.go
@@ -4,17 +4,16 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/thomasobenaus/sokar/test/integration/helper"
 	"github.com/thomasobenaus/sokar/test/integration/nomad"
 )
 
-func TestSimple(t *testing.T) {
-	testCase := "Simple"
-	sokarAddr := "http://localhost:11000"
-	nomadAddr := "http://localhost:4646"
+func TestScaleUp(t *testing.T) {
+	testCase := "ScaleUp"
+	sokarAddr := helper.SokarAddr
+	nomadAddr := helper.NomadAddr
 	jobName := "fail-service"
 
 	helper.PrintCheckPoint(testCase, "Start waiting for nomad (%s)....\n", nomadAddr)
@@ -48,39 +47,8 @@ func TestSimple(t *testing.T) {
 	helper.PrintCheckPoint(testCase, "Scale alert sent\n")
 
 	// ensure to disable the alert from firing
-	//defer helper.SendScaleAlert("AlertA", false)
-	NewJobAsserter(t, jobName, time.Millisecond*500, 30).AssertJobCount(34, jobName, d.GetJobCount)
+	defer helper.SendScaleAlert("AlertA", false)
 
-}
-
-type JobAsserter struct {
-	t        *testing.T
-	jobName  string
-	waitTime time.Duration
-	maxTries int
-}
-
-func NewJobAsserter(t *testing.T, jobName string, waitTime time.Duration, maxTries int) *JobAsserter {
-	return &JobAsserter{
-		t:        t,
-		jobName:  jobName,
-		waitTime: waitTime,
-		maxTries: maxTries,
-	}
-}
-
-func (ja *JobAsserter) AssertJobCount(expectedJobCount int, jobName string, obtainJobCountFunc func(jobName string) (int, error)) {
-	count := 0
-	for i := 0; i < ja.maxTries; i++ {
-		var err error
-		count, err = obtainJobCountFunc(ja.jobName)
-		require.NoError(ja.t, err, "Failed to obtain job count")
-		if count == expectedJobCount {
-			return // success case -> no assert, just return
-		}
-		ja.t.Logf("Jobcount not as expected (%d), but was %d. Recheck in %s\n", expectedJobCount, count, ja.waitTime)
-		time.Sleep(ja.waitTime)
-	}
-
-	assert.Failf(ja.t, "Jobcount invalid", "Jobcount is not %d as expected but was %d at last try (#tries=%d).", expectedJobCount, count, ja.maxTries)
+	helper.PrintCheckPoint(testCase, "Check if job was scaled to the expected count\n")
+	helper.NewJobAsserter(t, jobName, time.Millisecond*500, 50).AssertJobCount(3, jobName, d.GetJobCount)
 }
